requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
trafilatura==1.12.2
openai==1.51.0
python-dotenv==1.0.1
kokoro>=0.9.2
soundfile==0.12.1
numpy==1.26.4
httpx==0.27.2
python-multipart==0.0.12
slowapi==0.1.9
rumps==0.4.0
